requirements.txt
Django>=3.0.3,<3.1.0
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.8.4,<2.9.0

phonenumbers>=8.11.4
django-phonenumber-field>=4.0.0,<4.1.0

Pillow>=7.0.0,<7.1.0
qrcode>=6.1,<6.2

django-knox-rest>=4.1.0,<4.2.0

requests
